improve transaction tests
