+++ rskj-core/src/test/java/org/ethereum/TestUtils.java
@@ -19,9 +19,17 @@
 
 package org.ethereum;
 
-import co.rsk.core.Coin;
-import co.rsk.core.RskAddress;
-import co.rsk.crypto.Keccak256;
+import static junit.framework.TestCase.assertEquals;
+import static junit.framework.TestCase.assertNotNull;
+
+import java.io.File;
+import java.math.BigInteger;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Random;
+
 import org.apache.commons.lang3.StringUtils;
 import org.ethereum.core.Block;
 import org.ethereum.core.BlockFactory;
@@ -31,14 +39,9 @@ import org.ethereum.vm.DataWord;
 import org.mapdb.DB;
 import org.mapdb.DBMaker;
 
-import java.io.File;
-import java.math.BigInteger;
-import java.util.*;
-import java.util.function.Function;
-import java.util.function.Supplier;
-
-import static junit.framework.TestCase.*;
-import static org.ethereum.crypto.HashUtil.EMPTY_TRIE_HASH;
+import co.rsk.core.Coin;
+import co.rsk.core.RskAddress;
+import co.rsk.crypto.Keccak256;
 
 public final class TestUtils {
