+++ rskj-core/src/test/java/org/ethereum/util/TransactionFactoryHelper.java
@@ -1,9 +1,12 @@
 package org.ethereum.util;
 
+import co.rsk.core.RskAddress;
 import co.rsk.test.builders.AccountBuilder;
 import co.rsk.test.builders.TransactionBuilder;
 import org.ethereum.core.Account;
 import org.ethereum.core.Transaction;
+import org.ethereum.rpc.CallArguments;
+import org.ethereum.rpc.TypeConverter;
 
 import java.math.BigInteger;
 
@@ -11,6 +14,7 @@ import java.math.BigInteger;
  * Created by ajlopez on 28/02/2018.
  */
 public class TransactionFactoryHelper {
+
     public static Account createAccount(int naccount) {
         return new AccountBuilder().name("account" + naccount).build();
     }
@@ -80,4 +84,35 @@ public class TransactionFactoryHelper {
 
         return tx;
     }
+    
+	public static CallArguments createArguments(RskAddress sender, RskAddress receiver) {
+
+		// Simulation of the args handled in the sendTransaction call
+		CallArguments args = new CallArguments();
+		args.setFrom(sender.toJsonString());
+		args.setTo(receiver.toJsonString());
+		args.setGasLimit("0x76c0");
+		args.setGasPrice("0x9184e72a000");
+		args.setValue("0x186A0");
+		args.setNonce("0x01");
+
+		return args;
+	}
+
+	public static Transaction createTransaction(CallArguments args, byte chainId, Account senderAccount) {
+
+		// Transaction that is expected to be constructed WITH the gasLimit
+		Transaction tx = Transaction.builder()
+			.nonce(TypeConverter.stringNumberAsBigInt(args.getNonce()))
+			.gasPrice(TypeConverter.stringNumberAsBigInt(args.getGasPrice()))
+			.gasLimit(TypeConverter.stringNumberAsBigInt(args.getGasLimit()))
+			.destination(TypeConverter.stringHexToByteArray(args.getTo()))
+			.chainId(chainId)
+			.value(TypeConverter.stringNumberAsBigInt(args.getValue()))
+			.build();
+		tx.sign(senderAccount.getEcKey().getPrivKeyBytes());
+
+		return tx;
+	}
+    
 }
